+++ ChangeLog.rst
@@ -1,3 +1,13 @@
+UNRELEASED
+*******************
+
+Note worthy changes
+-------------------
+
+- Make the account enumeration prevention in ``BaseSignupForm`` friendlier to
+  inheritance.
+
+
 0.54.0 (2023-03-31)
 *******************